+++ tests/testthat/test-validate.R
@@ -9,6 +9,6 @@ test_that("Validation works", {
   validate_data(complex(rep(1, 2), c(1, -1)), validate_complex())
   validate_data(x = c('A', 'B'), validate_character(), validate_integer())
   validate_membership(c('A', 'B'))(c('A', 'C'))
-  validate_date(c('2023-01-01', '01'))
-  validate_time(c('23:59:59', '59'))
+  validate_date()(c('2023-01-01', '01'))
+  validate_time()(c('23:59:59', '59'))
 })